Display process updated, now correctly shows steps, km, mi

+++ 361Project/stepCounter.c
@@ -13,14 +13,14 @@
 #define THRESHOLD_ACCEL 384
 
 // takes x, y and z mean acceleration inputs and returns the norm, raw
-int16_t calculate_norm(vector3_t mean_acc)
+int32_t calculate_norm(vector3_t mean_acc)
 {
-    int16_t accel_norm = sqrt(mean_acc.x*mean_acc.x + mean_acc.y*mean_acc.y + mean_acc.z*mean_acc.z);
+    int32_t accel_norm = sqrt(mean_acc.x*mean_acc.x + mean_acc.y*mean_acc.y + mean_acc.z*mean_acc.z);
     return accel_norm;
 }
 
 // toggles flag if current value of the acceleration norm is greater than 1.5G
-int less_than_flag (int16_t current_norm)
+int less_than_flag (uint32_t current_norm)
 {
     int is_flagged;
     if (current_norm > THRESHOLD_ACCEL)
@@ -33,7 +33,7 @@ int less_than_flag (int16_t current_norm)
 }
 
 // increments total steps if the norm of the accelerations increases from less than to more than 1.5G
-int16_t step_increment (int current_steps, int current_flag, int prev_flag)
+int16_t step_increment (uint32_t current_steps, uint8_t current_flag, uint8_t prev_flag)
 {
     if (current_flag == 1 && prev_flag == 0)
     {
