+++ 361Project/stepCounter.h
@@ -17,12 +17,12 @@
 #include <math.h>
 
 // takes x, y and z mean acceleration inputs and returns the norm, raw
-int16_t calculate_norm(vector3_t mean_acc);
+int32_t calculate_norm(vector3_t mean_acc);
 
 // toggles flag if current value of the acceleration norm is greater than 1.5G
-int less_than_flag (int16_t current_norm);
+int less_than_flag (uint32_t current_norm);
 
 // increments total steps if the norm of the accelerations increases from less than to more than 1.5G
-int16_t step_increment (int current_steps, int current_flag, int prev_flag);
+int16_t step_increment (uint32_t current_steps, uint8_t current_flag, uint8_t prev_flag);
 
 #endif /* STEPCOUNTER_H_ */
